convert NotFounPage to typescript
